make sure the Postgres version works as well + test DB migrations

+++ src/data/src/entity/pam/authorized_keys.rs
@@ -97,7 +97,11 @@ VALUES ($1, $2, $3, $4, $5, $6)"#;
                 .execute(sql, params!(pam_uid, now, expires, typ, data, comment))
                 .await?;
         } else {
-            DB::pg_execute(sql, &[&pam_uid, &now, &expires, &typ, &data, &comment]).await?;
+            DB::pg_execute(
+                sql,
+                &[&(pam_uid as i64), &now, &expires, &typ, &data, &comment],
+            )
+            .await?;
         }
 
         Ok(())
@@ -126,7 +130,7 @@ SELECT * FROM ssh_auth_keys WHERE pam_uid = (
         let res = if is_hiqlite() {
             DB::hql().query_map(sql, params!(pam_uid)).await?
         } else {
-            DB::pg_query(sql, &[&pam_uid], 1).await?
+            DB::pg_query(sql, &[&(pam_uid as i64)], 1).await?
         };
 
         Ok(res)
@@ -140,7 +144,7 @@ SELECT * FROM ssh_auth_keys WHERE pam_uid = (
                 .query_map_one(sql, params!(pam_uid, ts_added))
                 .await?
         } else {
-            DB::pg_query_one(sql, &[&pam_uid, &ts_added]).await?
+            DB::pg_query_one(sql, &[&(pam_uid as i64), &ts_added]).await?
         };
 
         Ok(slf)
@@ -195,7 +199,7 @@ VALUES ($1, $2)
                 let now = Utc::now().timestamp();
                 DB::pg_txn_append(&txn, sql_blacklist, &[&hash, &now]).await?;
             }
-            DB::pg_txn_append(&txn, sql_delete, &[&self.pam_uid, &self.ts_added]).await?;
+            DB::pg_txn_append(&txn, sql_delete, &[&(self.pam_uid as i64), &self.ts_added]).await?;
 
             txn.commit().await?;
         }
